# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Code Challenge - Squid Game",
    "version": "1.0.0"
  },
  "schemes": ["http", "https"],
  "paths": {
...
    "/api/v1.0/death": {
      "get": {
        "description": "Get a list of all Deaths",
        "produces": ["application/json"],
        "tags": ["death"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Death"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a Death",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Death"
              }
            }
          }
        },
        "produces": ["application/json"],
        "tags": ["death"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "$ref": "#/definitions/Death"
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "put": {
        "description": "Update a Death",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Death"
              }
            }
          }
        },
        "produces": ["application/json"],
        "tags": ["death"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "$ref": "#/definitions/Death"
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      }
    },
    "/api/v1.0/death/{id}": {
      "get": {
        "description": "Get a Death by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Death ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "produces": ["application/json"],
        "tags": ["death"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "$ref": "#/definitions/Death"
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "Death": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "uid": {
          "type": "string",
          "format": "uuid"
        },
        "number": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}